Time: 1 ms (100.00%) | Memory: 42.9 MB (9.09%) - LeetSync

+++ 441-arranging-coins/arranging-coins.java
@@ -11,7 +11,7 @@ class Solution {
         return (int) end ;
      }
     public int arrangeCoins(int n) {
-        long m= (long) n;
+        long m=  n;
         return (mySqrt(8*m+1)-1)/2;
     }
 }
